renamed to ServerConstants
